Implement isolate-based frame processing; define interface and separate async/sync processing logic for improved performance and maintainability.
